requirements.txt
datasets==2.11.0
# transformers==4.27.4
adapter-transformers==3.2.0
numpy==1.24.2
torch==2.0.0